+++ CHANGELOG.md
@@ -32,6 +32,8 @@ the REPL ([#120](https://github.com/anmonteiro/lumo/issues/120)).
 - JavaScript completion only works in readline mode ([#103](https://github.com/anmonteiro/lumo/issues/103)).
 - Comments in the REPL cause reader exception errors ([#74](https://github.com/anmonteiro/lumo/issues/74)).
 - Fix error when calling `child_process.fork` ([#53](https://github.com/anmonteiro/lumo/issues/53)).
+- Fix a bug that would prevent TTY support on Windows. Lumo is now at feature-parity
+with every other supported platform.
 
 ## [1.3.0](https://github.com/anmonteiro/lumo/compare/1.2.0...1.3.0) (2017-04-02)